Change according to PR comments
