+++ nvtabular/visualization/statistics.py
@@ -1,0 +1,46 @@
+import base64
+import os
+
+STATS_FILE_NAME = "stats.pb"
+
+
+class DatasetCollectionStatistics:
+    HTML_TEMPLATE = """
+<script
+    src="https://cdnjs.cloudflare.com/ajax/libs/webcomponentsjs/1.3.3/webcomponents-lite.js">
+</script>
+<link rel="import"
+    href="https://raw.githubusercontent.com/PAIR-code/facets/1.0.0/facets-dist/facets-jupyter.html">
+<facets-overview id="elem"></facets-overview>
+<script>
+ document.querySelector("#elem").protoInput = "{protostr}";
+</script>"""
+
+    def __init__(self, dataset_feature_statistics_list) -> None:
+        super().__init__()
+        self.stats = dataset_feature_statistics_list
+
+    def display_overview(self):
+        from IPython.core.display import HTML, display
+
+        return display(HTML(self.to_html()))
+
+    def to_html(self):
+        protostr = self.to_proto_string(self.stats)
+        html = self.HTML_TEMPLATE.format(protostr=protostr)
+
+        return html
+
+    def save_to_html(self, output_dir, file_name="stats.html"):
+        with open(os.path.join(output_dir, file_name), "w") as html_file:
+            html_file.write(self.to_html())
+
+    def to_proto_string(self, inputs):
+        return base64.b64encode(inputs.SerializeToString()).decode("utf-8")
+
+    def save(self, output_dir, file_name=STATS_FILE_NAME):
+        out_path = os.path.join(output_dir, file_name)
+        with open(out_path, "wb") as f:
+            f.write(self.stats.SerializeToString())
+
+        self.save_to_html(output_dir)
